requirements.txt
pandas
yfinance
yagmail
requests
gspread
oauth2client
python-dotenv 